Rename ValueObserver -> Gauge
