--- sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/LongValueObserverSdk.java
@@ -1,54 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.sdk.metrics;
-
-import io.opentelemetry.api.metrics.DoubleGaugeBuilder;
-import io.opentelemetry.api.metrics.LongGaugeBuilder;
-import io.opentelemetry.api.metrics.ObservableLongMeasurement;
-import io.opentelemetry.sdk.metrics.common.InstrumentType;
-import io.opentelemetry.sdk.metrics.internal.state.MeterProviderSharedState;
-import io.opentelemetry.sdk.metrics.internal.state.MeterSharedState;
-import java.util.function.Consumer;
-
-final class LongValueObserverSdk {
-
-  private LongValueObserverSdk() {}
-
-  static final class Builder extends AbstractInstrumentBuilder<LongValueObserverSdk.Builder>
-      implements LongGaugeBuilder {
-
-    Builder(
-        MeterProviderSharedState meterProviderSharedState,
-        MeterSharedState meterSharedState,
-        String name) {
-      this(meterProviderSharedState, meterSharedState, name, "", "1");
-    }
-
-    Builder(
-        MeterProviderSharedState meterProviderSharedState,
-        MeterSharedState sharedState,
-        String name,
-        String description,
-        String unit) {
-      super(meterProviderSharedState, sharedState, name, description, unit);
-    }
-
-    @Override
-    protected Builder getThis() {
-      return this;
-    }
-
-    @Override
-    public DoubleGaugeBuilder ofDoubles() {
-      return swapBuilder(DoubleValueObserverSdk.Builder::new);
-    }
-
-    @Override
-    public void buildWithCallback(Consumer<ObservableLongMeasurement> callback) {
-      registerLongAsynchronousInstrument(InstrumentType.VALUE_OBSERVER, callback);
-    }
-  }
-}
